requirements.txt
# Telegram API
telethon>=1.36.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
newspaper3k>=0.2.8

# NLP and similarity
sentence-transformers>=2.3.0
transformers>=4.36.0
torch>=2.1.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.0
jsonschema>=4.20.0
